optimize resume retry
