+++ base/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.8.0
+
+- 优化分片上传进度计算，改为基于字节权重的进度
+- 修复分片上传任务生命周期管理问题
+- 修复分片上传断点续传恢复逻辑
+- 新增 `RESOURCE_READ_EXCEPTION` 错误类型
+- 新增 HTTP 客户端读写空闲超时配置
+- 修复 User-Agent 偶尔无效的问题
+
 ## 0.7.5
 
 - 文件进行第一次上传时，如果域名全部冻结，会随机选择一个域名进行上传
