+++ base/lib/src/storage/methods/put/by_part/cache_mixin.dart
@@ -10,11 +10,63 @@ mixin CacheMixin<T> on RequestTask<T> {
     await config.cacheProvider.removeItem(_cacheKey);
   }
 
-  Future<void> setCache(String data) async {
-    await config.cacheProvider.setItem(_cacheKey, data);
+  Future<void> setCache(String data, {int? expireAt}) async {
+    final cacheItem = _CacheItem(data, expireAt: expireAt ?? 0);
+    await config.cacheProvider.setItem(_cacheKey, cacheItem._toCacheString());
   }
 
   Future<String?> getCache() async {
-    return await config.cacheProvider.getItem(_cacheKey);
+    final cacheString = await config.cacheProvider.getItem(_cacheKey);
+    final cacheItem = _CacheItem.fromCacheString(cacheString);
+    if (cacheItem == null) {
+      return null;
+    }
+
+    if (cacheItem.isExpired) {
+      await config.cacheProvider.removeItem(_cacheKey);
+      return null;
+    }
+
+    return cacheItem.data;
+  }
+}
+
+class _CacheItem {
+  final String data;
+  // 单位：毫秒
+  final int expireAt;
+
+  _CacheItem(this.data, {this.expireAt = 0});
+
+  static _CacheItem? fromCacheString(String? cacheString) {
+    if (cacheString == null) {
+      return null;
+    }
+    try {
+      final cacheMap = jsonDecode(cacheString) as Map<String, dynamic>;
+      return _CacheItem(
+        cacheMap['data'] as String,
+        expireAt: cacheMap['expireAt'] as int,
+      );
+    } catch (e) {
+      // 解析失败说明缓存数据有问题，直接当做没有缓存
+      return _CacheItem('', expireAt: -1);
+    }
+  }
+
+  String _toCacheString() {
+    final cacheMap = {
+      'data': data,
+      'expireAt': expireAt,
+    };
+    return jsonEncode(cacheMap);
+  }
+
+  bool get isExpired {
+    if (expireAt == 0) {
+      return false;
+    }
+    final currentTime = DateTime.now().millisecondsSinceEpoch;
+    return currentTime >= expireAt;
   }
 }
