+++ base/lib/src/storage/methods/put/by_part/init_parts_task.dart
@@ -103,7 +103,14 @@ class InitPartsTask extends RequestTask<InitParts> with CacheMixin<InitParts> {
 
   @override
   Future<void> postReceive(data) async {
-    await setCache(json.encode(data.toJson()));
+    // 有效期往前推 3 小时，避免在上传过程中缓存过期导致的 uploadId 无效问题
+    final expireAt = (data.expireAt - 3 * 3600) * 1000;
+    if (expireAt > DateTime.now().millisecondsSinceEpoch) {
+      await setCache(
+        json.encode(data.toJson()),
+        expireAt: data.expireAt * 1000,
+      );
+    }
     await super.postReceive(data);
   }
 }
