+++ base/lib/src/storage/storage.dart
@@ -2,27 +2,26 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:path/path.dart' show basename;
+import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
 
-import 'config/config.dart';
 import 'methods/put/by_part/put_parts_task.dart';
 import 'methods/put/by_single/put_by_single_task.dart';
-import 'methods/put/put.dart';
 import 'resource/resource.dart';
-import 'task/task.dart';
+import 'task/request_task.dart';
+import 'task/task_manager.dart';
 
 export 'package:dio/dio.dart' show HttpClientAdapter;
-export '../client/http_client_adapter.dart';
 export 'error/error.dart';
 export 'methods/put/put.dart';
 export 'status/status.dart';
-export 'task/request_task.dart';
-export 'task/task.dart';
 export 'config/config.dart';
 
 /// 客户端
 class Storage {
   final Config config;
-  final TaskManager taskManager = TaskManager();
+
+  /// 任务管理器，负责管理上传任务的执行
+  final TaskManager _taskManager = TaskManager();
 
   Storage({Config? config}) : config = config ?? Config();
 
@@ -59,7 +58,14 @@ class Storage {
       );
     }
 
-    taskManager.addTask(task);
+    if (_taskManager.hasTask(task.taskID())) {
+      throw StorageError(
+        type: StorageErrorType.IN_PROGRESS,
+        message: '$resource is already in upload queue',
+      );
+    }
+
+    _taskManager.addTask(task);
 
     return task.future;
   }
@@ -97,7 +103,14 @@ class Storage {
       );
     }
 
-    taskManager.addTask(task);
+    if (_taskManager.hasTask(task.taskID())) {
+      throw StorageError(
+        type: StorageErrorType.IN_PROGRESS,
+        message: '$resource is already in upload queue',
+      );
+    }
+
+    _taskManager.addTask(task);
 
     return task.future;
   }
