+++ base/lib/src/storage/task/task_manager.dart
@@ -3,26 +3,34 @@ import 'package:meta/meta.dart';
 import 'task.dart';
 
 class TaskManager {
-  final List<Task> workingTasks = [];
+  final List<Task> _workingTasks = [];
+
+  /// 判断是否有 taskID 对应的 Task 正在运行
+  bool hasTask(String taskID) {
+    return _workingTasks.any((task) => task.taskID() == taskID);
+  }
 
   /// 添加一个 [Task]
   ///
   /// 被添加的 [task] 会被立即执行 [createTask]
+  /// task 执行完就会从管理器中移除
   @mustCallSuper
   Future<void> addTask(Task task) async {
+    _workingTasks.add(task);
     try {
       await task.preStart();
-      workingTasks.add(task);
       final taskFuture = task.createTask();
       await task.postStart();
 
       final result = await taskFuture;
+      await removeTask(task);
       await task.postReceive(result);
     } catch (error) {
       if (task.showRetry(error)) {
         await task.postError(error, complete: false);
         await restartTask(task);
       } else {
+        await removeTask(task);
         await task.postError(error, complete: true);
       }
     }
@@ -30,7 +38,7 @@ class TaskManager {
 
   @mustCallSuper
   Future<void> removeTask(Task task) async {
-    workingTasks.remove(task);
+    _workingTasks.remove(task);
   }
 
   @mustCallSuper
@@ -38,16 +46,17 @@ class TaskManager {
     try {
       await task.preRestart();
       final taskFuture = task.createTask();
-      workingTasks.add(task);
       await task.postRestart();
 
       final result = await taskFuture;
+      await removeTask(task);
       await task.postReceive(result);
     } catch (error) {
       if (task.showRetry(error)) {
         await task.postError(error, complete: false);
         await restartTask(task);
       } else {
+        await removeTask(task);
         await task.postError(error, complete: true);
       }
     }
@@ -55,11 +64,11 @@ class TaskManager {
 
   /// 返回当前运行中的 [Task]
   List<Task<dynamic>> getTasks() {
-    return workingTasks;
+    return _workingTasks;
   }
 
   /// 查找类型符合 [T] 的 [Task]
   List<T> getTasksByType<T extends Task<dynamic>>() {
-    return workingTasks.whereType<T>().toList();
+    return _workingTasks.whereType<T>().toList();
   }
 }
