+++ base/test/put/put_by_parts/put_bytes_test.dart
@@ -1,5 +1,6 @@
 @Timeout(Duration(seconds: 60))
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
+import 'package:qiniu_sdk_base/src/storage/task/task_manager.dart';
 import 'package:qiniu_sdk_base/src/storage/methods/put/by_part/put_parts_task.dart';
 import 'package:qiniu_sdk_base/src/storage/resource/resource.dart';
 import 'package:test/test.dart';
@@ -71,8 +72,8 @@ void main() {
         ),
       );
       expect(response, isA<PutResponse>());
-      // 2 片分片所以 2 次
-      expect(callnumber, 2);
+      // 分票进度回调要多次调用，进度尽可能连续
+      expect(callnumber > 10, true);
 
       pcb.testAll();
 
@@ -314,9 +315,8 @@ void main() {
         resource: resource,
         key: key,
       );
-
-      storage.taskManager.addTask(task);
-
+      final tmpTaskManager = TaskManager();
+      tmpTaskManager.addTask(task);
       await task.future;
 
       final putController = PutController();
