+++ base/test/put/put_by_parts/put_file_test.dart
@@ -4,6 +4,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:qiniu_sdk_base/qiniu_sdk_base.dart';
+import 'package:qiniu_sdk_base/src/storage/task/task_manager.dart';
 import 'package:qiniu_sdk_base/src/storage/methods/put/by_part/put_parts_task.dart';
 import 'package:qiniu_sdk_base/src/storage/resource/resource.dart';
 import 'package:test/test.dart';
@@ -76,7 +77,7 @@ void main() {
       );
       expect(response, isA<PutResponse>());
       // 2 片分片所以 2 次
-      expect(callnumber, 2);
+      expect(callnumber > 10, true);
 
       pcb.testAll();
 
@@ -281,8 +282,8 @@ void main() {
         key: key,
       );
 
-      storage.taskManager.addTask(task);
-
+      final tmpTaskManager = TaskManager();
+      tmpTaskManager.addTask(task);
       await task.future;
 
       final putController = PutController();
